fix: resolve critical type checking errors in refactored codebase
- Fix function signature mismatches for extract_prd_summary
- Add None path validation for all Path() constructions
- Fix possibly unbound variable errors in ai_code_review_enhanced.py
- Remove duplicate TaskData declarations and fix import order
- Remove unused imports (ClaudeMemoryFile, CursorRule, parse_github_pr_url)
- Update test imports to match new module structure

Reduced pyright errors from 56 to 30 across the codebase.
